created a shopping list program
so far the program adds items to an unordered list, throught the use
of a simple input field and a button, using and event listener to
listen for a click on the button and then adding the input value to
the list. Next I will add a  event listener to the delete button so each list item
can be deleted individually.